[ADDED] move anim to s2
Former-commit-id: 365b3626c4c38005c9a8d00378ecc6bb0d914be4

+++ sprite2/source/sprites/BezierShape.cpp
@@ -42,7 +42,7 @@ void BezierShape::UpdatePolyline()
 	const int num = std::max(20, (int)(sm::dis_pos_to_pos(m_control_nodes[0], m_control_nodes[3]) / 10));
 	float dt = 1.0f / (num - 1);
 	std::vector<sm::vec2> vertices(num);
-	for (size_t i = 0; i < num; ++i) {
+	for (int i = 0; i < num; ++i) {
 		vertices[i] = PointOnCubicBezier(i * dt);
 	}
